add post,edit, delete

+++ controllers/adminController.js
@@ -7,74 +7,9 @@ const Blog = require("../models/Blog");
 var path = require("path");
 
 const { fileFilter } = require("../utils/multer");
+const util = require("../utils/helpers");
 
-exports.getDashboard = async (req, res) => {
-  const page = +req.query.page || 1;
-  const postPerPage = process.env.PAGE_CONTENT;
-
-  try {
-    const numberOfPosts = await Blog.find({
-      user: req.user._id,
-    }).countDocuments();
-
-    const blogs = await Blog.find({ user: req.user.id })
-      .sort({
-        createdAt: "desc",
-      })
-      .skip((page - 1) * postPerPage)
-      .limit(postPerPage);
-
-    res.render("private/blogs", {
-      pageTitle: "بخش مدیریت | داشبورد",
-      path: "/dashboard",
-      layout: "./layouts/dashlayout",
-      fullname: req.user.fullname,
-      blogs,
-      currentPage: page,
-      nextPage: page + 1,
-      previousPage: page - 1,
-      hasNextPage: postPerPage * page < numberOfPosts,
-      hasPreviousPage: page > 1,
-      lastPage: Math.ceil(numberOfPosts / postPerPage),
-      postPerPage,
-    });
-  } catch (error) {
-    console.log(error);
-  }
-};
-
-exports.getAddPost = (req, res) => {
-  res.render("private/addPost", {
-    pageTitle: "بخش مدیریت | ساخت پست جدید",
-    path: "/dashboard/add-post",
-    layout: "./layouts/dashLayout",
-    fullname: req.user.fullname,
-  });
-};
-
-exports.getEditPost = async (req, res) => {
-  const post = await Blog.findOne({
-    _id: req.params.id,
-  });
-
-  if (!post) {
-    return res.redirect("errors/404");
-  }
-
-  if (post.user.toString() != req.user._id) {
-    return res.redirect("/dashboard");
-  } else {
-    res.render("private/editPost", {
-      pageTitle: "بخش مدیریت | ویرایش پست",
-      path: "/dashboard/edit-post",
-      layout: "./layouts/dashLayout",
-      fullname: req.user.fullname,
-      post,
-    });
-  }
-};
-
-exports.editPost = async (req, res) => {
+exports.editPost = async (req, res, next) => {
   const errorArr = [];
 
   const post = await Blog.findOne({ _id: req.params.id });
@@ -86,10 +21,14 @@ exports.editPost = async (req, res) => {
     await Blog.postValidation(req.body);
 
     if (!post) {
-      return res.redirect("errors/404");
+      const error = new Error("پست یافت نشد");
+      error.statusCode = 400;
+      throw error;
     }
-    if (post.user.toString() != req.user._id) {
-      return res.redirect("/dashboard");
+    if (post.user.toString() != req.userId) {
+      const error = new Error("شما مجوز ویرایش این پست را ندارید");
+      error.statusCode = 400;
+      throw error;
     } else {
       const { title, status, body } = req.body;
 
@@ -118,7 +57,7 @@ exports.editPost = async (req, res) => {
 
       await post.save();
 
-      return res.redirect("/dashboard");
+      return res.status(200).json({ message: "پست با موفقیت ویرایش گردید" });
     }
   } catch (err) {
     console.log(err);
@@ -129,21 +68,20 @@ exports.editPost = async (req, res) => {
         message: e.message,
       });
     });
-    res.render("private/editPost", {
-      pageTitle: "بخش مدیریت | ویرایش پست",
-      path: "/dashboard/edit-post",
-      layout: "./layouts/dashLayout",
-      fullname: req.user.fullname,
-      errors: errorArr,
-      post,
-    });
+    res.status(400).json({ errorArr });
   }
-  res.redirect("/dashboard");
 };
 
-exports.deletePost = async (req, res) => {
+exports.deletePost = async (req, res, next) => {
   try {
     const post = await Blog.findOne({ _id: req.params.id });
+
+    if (!post) {
+      const error = new Error("پست یافت نشد");
+      error.statusCode = 400;
+      throw error;
+    }
+
     fs.unlink(
       `${appRoot}/public/uploads/thumbnails/${post.thumbnail}`,
       async (err) => {
@@ -152,22 +90,32 @@ exports.deletePost = async (req, res) => {
       }
     );
 
-    res.redirect("/dashboard");
+    res.status(200).json({ message: "پست با موفقیت حذف گردید" });
   } catch (err) {
     console.log(err);
-    res.render("errors/500");
+    next(err);
   }
 };
 
 exports.createPost = async (req, res) => {
   const errorArr = [];
   const thumbnail = req.files ? req.files.thumbnail : {};
-  const fileName = `${util.replaceAll(uuid(), "-", "")}${path.extname(
-    thumbnail.name
-  )}`;
-  const uploadPath = `${appRoot}/public/uploads/thumbnails/${fileName}`;
-
   try {
+    if (!req.files) {
+      errorArr.push({
+        name: "image",
+        message: "تصویر الزامی می باشد",
+      });
+
+      throw error;
+    }
+
+    const fileName = `${util.replaceAll(uuid(), "-", "")}${path.extname(
+      thumbnail.name
+    )}`;
+
+    const uploadPath = `${appRoot}/public/uploads/thumbnails/${fileName}`;
+
     req.body = { ...req.body, thumbnail };
 
     await Blog.postValidation(req.body);
@@ -177,10 +125,18 @@ exports.createPost = async (req, res) => {
       .toFile(uploadPath)
       .catch((err) => console.log(err));
 
-    await Blog.create({ ...req.body, user: req.user.id, thumbnail: fileName });
+    const postCreated = await Blog.create({
+      ...req.body,
+      user: req.userId,
+      thumbnail: fileName,
+    });
+
+    res.status(200).json({
+      _id: postCreated._id.toString(),
+      message: "مطلب با موفقیت ذخیره گردید",
+    });
   } catch (err) {
     console.log(err);
-    //get500(req, res);
 
     if (err.inner) {
       err.inner.forEach((e) => {
@@ -190,84 +146,41 @@ exports.createPost = async (req, res) => {
         });
       });
     }
-    res.render("private/addPost", {
-      pageTitle: "بخش مدیریت | ساخت پست جدید",
-      path: "/dashboard/add-post",
-      layout: "./layouts/dashLayout",
-      fullname: req.user.fullname,
-      errors: errorArr,
-    });
+    return res.status(400).json({ errorArr });
   }
-  res.redirect("/dashboard");
 };
 
 exports.uploadImage = (req, res) => {
   const upload = multer({
     limits: { fileSize: 4000000 },
-    // dest: "uploads/",
-    // storage: storage,
     fileFilter: fileFilter,
   }).single("image");
 
   upload(req, res, async (err) => {
     if (err) {
       if (err.code === "LIMIT_FILE_SIZE") {
-        return res
-          .status(400)
-          .send("حجم عکس ارسالی نباید بیشتر از 4 مگابایت باشد");
+        return res.status(422).json({
+          error: "حجم عکس ارسالی نباید بیشتر از 4 مگابایت باشد",
+        });
       }
-      res.send(err);
+      res.status(400).json({ error: err });
     } else {
-      if (req.file) {
-        const filename = `${uuid()}_${req.file.originalname}`;
-        await sharp(req.file.buffer)
+      if (req.files) {
+        const fileName = `${shortId.generate()}_${req.files.image.name}`;
+        await sharp(req.files.image.data)
           .jpeg({
             quality: 60,
           })
-          .toFile(`./public/uploads/${filename}`)
+          .toFile(`./public/uploads/${fileName}`)
           .catch((err) => console.log(err));
-        res.status(200).send(`http://localhost:3000/uploads/${filename}`);
+        res.status(200).json({
+          image: `http://localhost:3000/uploads/${fileName}`,
+        });
       } else {
-        res.send("جهت آپلود باید عکسی انتخاب کنید");
+        res.status(400).json({
+          error: "جهت آپلود باید عکسی انتخاب کنید",
+        });
       }
     }
   });
-};
-
-exports.handleDashSearch = async (req, res) => {
-  const page = +req.query.page || 1;
-  const postPerPage = 2;
-
-  try {
-    const numberOfPosts = await Blog.find({
-      user: req.user._id,
-      //$text: { $search: req.body.search },
-      title: { $regex: ".*" + req.body.search + ".*" },
-    }).countDocuments();
-    const blogs = await Blog.find({
-      user: req.user.id,
-      //$text: { $search: req.body.search },
-      title: { $regex: ".*" + req.body.search + ".*" },
-    })
-      .skip((page - 1) * postPerPage)
-      .limit(postPerPage);
-
-    res.render("private/blogs", {
-      pageTitle: "بخش مدیریت | داشبورد",
-      path: "/dashboard",
-      layout: "./layouts/dashLayout",
-      fullname: req.user.fullname,
-      blogs,
-      currentPage: page,
-      nextPage: page + 1,
-      previousPage: page - 1,
-      hasNextPage: postPerPage * page < numberOfPosts,
-      hasPreviousPage: page > 1,
-      lastPage: Math.ceil(numberOfPosts / postPerPage),
-      postPerPage,
-    });
-  } catch (err) {
-    console.log(err);
-    get500(req, res);
-  }
 };
